Спринт 16.
исправлен плеер и поиск
