--- app/src/main/java/com/practicum/playlistmaker_ver2/player/domain/api/PlayerInteractor.kt
@@ -1,16 +1,0 @@
-package com.practicum.playlistmaker_ver2.player.domain.api
-
-
-import com.practicum.playlistmaker_ver2.player.ui.models.PlayerState
-
-
-interface PlayerInteractor {
-
-    fun playPause()
-
-    fun release()
-    fun setTrackUrl(trackUrl: String)
-    fun setStateChangeListener(listener: (PlayerState, Long, String?) -> Unit)
-}
-
-
